+++ src/scripts/main.js
@@ -1,5 +1,6 @@
 import { getEvents } from "./events/eventsProvider.js";
 import { EventsList } from "./events/EventList.js";
 import "./events/EventDialogButton.js"
+import "./events/EventForm.js"
 
 getEvents().then(EventsList)
